Fixed missing cast for literal `1` to UINT64

+++ Source/inv.cpp
@@ -1852,7 +1852,7 @@ void __fastcall RemoveInvItem(int pnum, int iv)
 		{
 			if ( !(
 				*(UINT64 *)&plr[pnum]._pScrlSpells // TODO: remove cast when pScrlSpells is converted to UINT64
-				& (1 << (plr[pnum]._pRSpell - 1))) )
+				& ((UINT64)1 << (plr[pnum]._pRSpell - 1))) )
 			{
 				plr[pnum]._pRSpell = SPL_INVALID;
 			}
@@ -1874,7 +1874,7 @@ void __fastcall RemoveSpdBarItem(int pnum, int iv)
 		{
 			if ( !(
 				*(UINT64 *)&plr[pnum]._pScrlSpells // TODO: remove cast when pScrlSpells is converted to UINT64
-				& (1 << (plr[pnum]._pRSpell - 1))) )
+				& ((UINT64)1 << (plr[pnum]._pRSpell - 1))) )
 			{
 				plr[pnum]._pRSpell = SPL_INVALID;
 			}
